Allow RTS calls prior to initial route table load
To better support heartbeat responses it is necessary to
allow RTS messages to be sent prior to the loading of
the initial route table. This change implements this and
has the side effect of also allowing wormhold sends prior
to a route table arrival (which is ok since no routing
is needed for wormhole traffic).

Issue-ID: RIC-589

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I302bdd0f274a7c5bc95f3f382e3374b21662ab8c