Bump version to deal with CI bug
The CI jobs failed to move all packages on the first go and
is now unable to move the remainter packages. This change
is solely to bump the version number to get the CI jobs to
work and push packages to release.  The version is now 4.5.1
and has NO code changes.

Issue-ID: RIC-732

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I4ed87731ad038aabb63509f9a079cf761705024d