Add ability to save route table updates to disk
Updates from the route table generator are now saved into a disk
file using the new RMR_STASH_RT environment variable, or adding a .stash
extension to the vaue of RMR_SEED_RT if the stash variable is
not present. If neither variable is present, then no information
is saved.  This is intended to make debugging easier; for non-RMR
developers reading the dump output is not straight forward.

Issue-ID: RIC-774

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I79d3adf8c3c291f82f54849a3be94648fb86aab1