Fix RMR routing statistic data printout crash
If a routing table update happens while traversing entries to print
routing statistics, then it will cause RMR to crash. This
crash is generated due to an unsafe reference used to print
the rt statistics. This change fixes this routing statistic
printout crash.

This change also fixes a related problem of read-access to ref count
in routing table that is not protected by a mutex.

Issue-ID: RIC-872

Signed-off-by: Alexandre Huff <alexandrehuff@utfpr.edu.br>
Change-Id: I03b0cb79a62e7fbdb1eb08c2006894fab4ebf5cd
Signed-off-by: Alexandre Huff <alexandrehuff@utfpr.edu.br>